separate main endpoint

+++ src/schemas/employee.rs
@@ -1,14 +1,15 @@
 use serde::Serialize;
 use mongodb::{ Database, Collection };
+use juniper::{ GraphQLObject };
 
-#[derive(Serialize)]
+#[derive(Serialize, GraphQLObject)]
 pub struct EmployeeSchema {
     pub ID: String,
     pub name: String,
     pub email : String,
     pub cellnumber : String,
     pub profilePicUrl : String,
-    pub monthlyWage: i16,
+    pub monthlyWage: i32,
     pub lastWageDay: i32,
     pub dayBeginningVacations : i32,
     pub lengthVacations : i32,
